feat(render): 3d transform & shaders

+++ include/granite/vector.hpp
@@ -5,7 +5,7 @@ namespace gr {
 class Vec2 {
 public:
     Vec2() : x(0), y(0) {}
-    Vec2(int xv, int yv) : x(xv), y(yv) {}
+    Vec2(float xv, float yv) : x(xv), y(yv) {}
 
     float x, y;
 };
@@ -13,7 +13,7 @@ public:
 class Vec3 {
 public:
     Vec3() : x(0), y(0), z(0) {}
-    Vec3(int xv, int yv, int zv) : x(xv), y(yv), z(zv) {}
+    Vec3(float xv, float yv, float zv) : x(xv), y(yv), z(zv) {}
 
     float x, y, z;
 };
